Create game.cpp
The player moves a rocket and try to kill all the UFOs (move: A, D, W, S; shoot: J, L, I, K). All UFOs behave randomly across the terminal and shoot every 2 seconds. The are Conway game cells that move across the game, if the rocket touches them they become purple and block UFO shoots, if any of the UFOs does so, they turn orange and block rocket shoots. A threads and mutex approach is planned to optimize the movement of all components of the game.